``` feat(iotws): 实现WebSocket负载均衡功能
- 修改Redis键值操作逻辑，从HGetAll改为Keys配合GetDel方式获取连接数
- 添加分布式锁机制确保数据一致性
- 重构负载均衡算法，增加服务器间连接数分配逻辑
- 新增WsBalancedPublish结构体用于传递负载均衡数据
- 调整定时任务执行频率从每秒改为每5分钟
- 更新缓存键命名规则，统一使用发布订阅模式
- 修复连接数统计类型转换错误问题
```

+++ app/iotws/cmd/internal/svc/connection.go
@@ -38,9 +38,9 @@ func (c *connection) GetNumber() int64 {
 	return c.count
 }
 
-func (c *connection) GetConnByCount(count int) ([]any, bool) {
+func (c *connection) GetConnByCount(count int64) ([]any, bool) {
 	conns := make([]any, count)
-	i := 0
+	i := int64(0)
 	c.soketMap.Range(func(key, conn any) bool {
 		i++
 		if i >= count {
